+++ packages/eyes-selenium/test/coverage/custom/TestSpecialCharacters.spec.js
@@ -28,6 +28,6 @@ describe(appName, () => {
     await webDriver.get('https://applitools.github.io/demo/TestPages/SpecialCharacters/index.html')
     await eyes.check('Test Special Characters', Target.window().fully())
     await eyes.close()
-    let allResults = await runner.getAllTestResults()
+    await runner.getAllTestResults()
   })
 })