+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -3,6 +3,27 @@
 ## Unreleased
 
 
+   ## 10.1.2 - 2020/5/13
+
+- added retry for requests returning with a status code of 503   
+
+## 10.1.1 - 2020/5/11
+
+- updated to @applitools/eyes-common@3.23.1
+
+## 10.1.0 - 2020/5/11
+
+- add iosDeviceInfo support
+- updated to @applitools/eyes-common@3.23.0
+
+## 10.0.2 - 2020/5/11
+
+- updated to @applitools/eyes-common@3.22.2
+
+## 10.0.1 - 2020/5/4
+
+- added resource info on putRender Error
+
 ## 10.0.0 - 2020/4/30
 
 - consolidate classes into core (no changes from 9.3.0)