+++ packages/visual-grid-client/CHANGELOG.md
@@ -3,6 +3,31 @@
 ## Unreleased
 
 
+## 14.0.1 - 2020/5/17
+
+- using "url" instead of "source" ad referer header and eyes test source.
+
+ ## 14.0.0 - 2020/5/17
+
+- BREAKING CHANGE - removed "referer" parameter from checkWindow, now using "source" 
+
+## 13.8.1 - 2020/5/13
+
+- updated to @applitools/eyes-sdk-core@10.1.2
+
+## 13.8.0 - 2020/5/11
+
+- updated to @applitools/eyes-sdk-core@10.1.1
+- add iosDeviceInfo support
+
+## 13.7.7 - 2020/5/7
+
+- updated to @applitools/dom-snapshot@3.5.2
+
+## 13.7.6 - 2020/5/4
+
+- updated to @applitools/eyes-sdk-core@10.0.1
+
 ## 13.7.5 - 2020/4/30
 
 - updated to @applitools/eyes-sdk-core@10.0.0